update system status boxes
