+++ app/(gcforms)/[locale]/(form administration)/form-builder/[id]/responses/[statusFilter]/components/SystemStatus/ResponsesOkay.tsx
@@ -1,0 +1,3 @@
+export const ResponsesOkay = () => {
+  return <div>Responses Okay</div>;
+};
